Moved drush updates into separate script. d7_make now calls d7_init if needed. Updated .gitignore.

+++ files/d7_make.sh
@@ -14,6 +14,12 @@ else
   exit 1;
 fi
 
+## Init site if it doesn't exist
+if [[ ! -e $SITEPATH ]]; then
+    sudo d7_init.sh $SITEPATH || exit 1;
+fi
+
+
 ## Delete build dir if it's there
 sudo -u apache rm -rf $SITEPATH/drupal_build
 
@@ -55,17 +61,5 @@ sudo rm -rf $SITEPATH/drupal_bak
 mv $SITEPATH/drupal $SITEPATH/drupal_bak
 mv $SITEPATH/drupal_build $SITEPATH/drupal
 
-## Enable update manager.
-sudo -u apache drush -y en update -r $SITEPATH/drupal || exit 1;
-
-## Apply security updates.
-sudo -u apache drush up -y --security-only -r $SITEPATH/drupal || exit 1;
-
-## Disable update manager; no need to leave it phoning home.
-sudo -u apache drush -y dis update -r $SITEPATH/drupal || exit 1;
-
-## Clear the caches
-sudo -u apache drush -y cc all -r $SITEPATH/drupal || exit 1;
-
-## Avoid a known performance-crusher in our environment
-sudo -u apache drush eval 'variable_set('drupal_http_request_fails', 0)' -r $SITEPATH/drupal || exit 1;
+## Apply security updates and clear caches.
+sudo d7_update.sh $SITEPATH || exit 1;
